requirements.txt
chromadb
sentence-transformers
matplotlib
PyPDF2
pillow
streamlit
requests